Optimize winit integration by removing a full clear and blit per frame
- EguiSoftwareRender doesn't hold canvas anymore, a new EguiSoftwareRenderCanvas does it
- EguiSoftwareRender::render allows to use cacheing algorithms without requiring a dedicated intermediate buffer (ie. canvas)
- EguiSoftwareRenderCanvas::render does what was done before
- New cacheing mode that doesn't store primitives render in memory:
  MeshCacheing: Cache primitives meshes, redraw primitives in the smallest changed bbox
- EguiSoftwareRender::render now output a bbox to reduce presentation cost when possible

+++ src/raster/tri.rs
@@ -59,8 +59,8 @@ pub fn draw_tri<const SUBPIX_BITS: i32>(
             if vert_uvs_vary {
                 vert_uv_stepper.attr += vert_uv_stepper.step_x * start as f32;
             }
-            let ss_start = (ss_min.x + start) as usize;
-            let ss_end = (ss_min.x + end) as usize;
+            let ss_start = (ss_min.x + start) as u32;
+            let ss_end = (ss_min.x + end) as u32;
 
             if simd && alpha_blend && !vert_uvs_vary {
                 #[cfg(any(target_arch = "x86_64", target_arch = "aarch64"))]
@@ -70,7 +70,7 @@ pub fn draw_tri<const SUBPIX_BITS: i32>(
                     #[cfg(target_arch = "x86_64")]
                     use crate::color_sse41::*;
 
-                    let dst = buffer.get_mut_span(ss_start, ss_end, ss_y as usize);
+                    let dst = buffer.get_mut_span(ss_start, ss_end, ss_y as u32);
                     if vert_col_vary {
                         #[cfg(target_arch = "x86_64")]
                         unsafe {
@@ -126,7 +126,7 @@ pub fn draw_tri<const SUBPIX_BITS: i32>(
                     } else {
                         draw.const_tri_color_u8x4
                     };
-                    let pixel = buffer.get_mut(ss_x as usize, ss_y as usize);
+                    let pixel = buffer.get_mut(ss_x as u32, ss_y as u32);
                     *pixel = if alpha_blend {
                         egui_blend_u8(src, *pixel)
                     } else {
